Crate spherro

Re-exports

pub use initializer::Config;

Modules

accelerators

fetcher

force

initializer
Question 7: What does the pub modifier do anyways?

kernel

particle

universe

util
Question 0:

Structs

Fetcher
Fetches data from the universe into a buffer for wasm to read. The point of this is to separate the Universe's concern from the data format needed by the client

Force

Particle

Universe